MINIHACK-HVE/requirements.txt
fastapi
uvicorn
python-dotenv
requests
pydantic
httpx
